Re-execute the bell file before each model execution.
That way, bell file definition can be used in the model.
This is done by passing a supplementary parameter 'bell' to the interpreter.
When bell is true we execute the bell file for building bell info:
events, relations and order definitions are processed,
When bell is false, we execute a cat file,events, relations and order definitions are ignored.

Also cleand order checking a bit




git-svn-id: svn+ssh://scm.cs.ucl.ac.uk/cs/research/pplv/virginia/repos/herd/herd@91 ca349960-8758-41ef-81b0-0e8c0ef917f4