Refactoring of bell execution and exploitation, I guess that the new organisation is clearer....
There are some semantical changes: the following two emum
declarations are special:

enum scopes = 'tag1 || ... 'tagN

Scope relations will be tag1 ... tagN, a primitive function tag2scopes
takes 'tagI as argument and returns the corresponding relation.
It is an error to define several enum scopes.

enum regions = 'tag1 || .... || 'tagN
Define the regions,  there will be one event per region depending
upon the region of the events location.
It is an error to define several enum regions.

The definition of function narrower is also special, the system deduces
the scope hierarchy from it.

The events construct is as before (except that I got rid of
the special declaration for regions)



git-svn-id: svn+ssh://scm.cs.ucl.ac.uk/cs/research/pplv/virginia/repos/herd/herd@108 ca349960-8758-41ef-81b0-0e8c0ef917f4