Change the inner working of the interpret for handling reification.
The value type (V.v) is now a polymorphic type ('a V.v),
where 'a is the type of accumulated results (last arg of the
exported interpret function). This type parameter contaminates env and st.

As a counter-measure, to hide this type parameter from the outside world,
the new types init_env and st_out are introduced.

The exported type of the interpreter now is:

      val interpret :
          S.test ->
            ks ->
              init_env ->
                S.rel_pp Lazy.t ->
                  (st_out -> 'a -> 'a) -> 'a -> 'a




git-svn-id: svn+ssh://scm.cs.ucl.ac.uk/cs/research/pplv/virginia/repos/herd/herd@203 ca349960-8758-41ef-81b0-0e8c0ef917f4