herd: Update atomic-ordered-before definition in aarch64.cat
Following late feedback from implementers, the definition of
atomic-ordered-before has been changed slightly in the ARMv8
architecture so that early forwarding from a Store-Exclusive to a plain
load (i.e. one without Acquire semantics) is permitted to occur, but
in the case that a Store-Exclusive is read by a load with Acquire
semantics, then the Store-Exclusive must be ordered before the load.

This change updates the aarch64.cat file accordingly.

Signed-off-by: Will Deacon <will.deacon@arm.com>