jingle: Add gen_theme to generate a theme file
gen_theme is introduced to generate the theme file with a map file and
a call file. The map file defines basic rules and functions. The call
file use the basic rules and functions to generate new rules. The
format of each call in call file is like:

"<left part>" -> "@<function name> <pattern>";

gen_theme will find if any rule's left part matches the pattern, if so
do the find-and-replace with the function.

Currently gen_theme only output the new rule set to stdout.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>