gen_theme: Allow to try applying the functions multiple times
Currently in gen_theme, we only support applying each call once in the
call file, however, in order to support rule generation sequence such as
relaxed -> release -> acq_release -> fully ordered, we need to be able
to apply functions on unresolved call terms with updated environment
multiple times.

So this patch changes the logic of gen_theme: it will evaluate on the
calls that could be resolved first, and update the environment and try
the unresolved calls again and so forth. A new option is added to limit
the loops for trying to resolve calls, the default value is 5.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>