jingle/jingle: Make fresh_label and label env global
fresh_label and label env used to be local to find_label, however during
the translation process, there are many instances for find_label,
moreover, for one substitution, we use one find_label for Label and
another find_label for branching instruction translation. This makes we
either only generate one label (namely 'lbl1') for all the labels, or we
cannot generate the same label for branching instruction and the target
label of branching.

So we make them global to the whole translation, so that we can generate
the correct labels.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>