jingle/jingle: Reset label_env for each substitution
In previous commit, we make label_env global, so that branching
instructions and targeting labels can generate the same label, however
this introduces another problem: we generate the same label for all the
label with the same name in translation rules. It may be OK in herd can
handle labels correctly, however, it could not unfortunately ;-(

So we reset label_env for each substitution, in this way, label
generations is kinda local to each substitution, and we generate
different names for the same label in different substitutions.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>