joker extremity to relaxations in diy
This patch adds a new feature to diy, called "joker extremities".  In a diy
cycle, a node can now be, in addition to a read (R) or a write (W), a joker
(J). A joker can link with any other extremity - this allows to insert
instructions that are neither store or load.

Signed-off-by: Jade Alglave <jade.alglave@arm.com>