jingle: Make expl_instr in one instanciate_with share the same reg env
In the translation, one instanciate_with is corrsponding to one
translation rule, so that (temporary) variables in one instanciate_with
should share the same register binding (i.e. the variables with the same
name should be mapped into the same register).

We can archieve this by introducing a shared reg_env in
instanciate_with, and make expl_instr to take a ref to it.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>