jingle: gen_theme: Support func sequence (#15)
"Func sequence" is a series of functions concatenated by vertical bars
"|" (or pipes). It provides a way to compose functions in defined in map
file to get another function. For example:

	"acq_rel" : "acquire | release"

, this defines a new function named "acq_rel", whose behavior is simply
doing a "release" on a rule after an "acquire", IOW, first apply
"acquire" function to a rule A, get a intermediate rule B, and then
applying "release" function to rule B, get a final rule C.

For example, if we define two funcs:

	"release" : "store:STR" -> "store:STLXR"
	"acquire" : "load:LDXR" -> "load:LDAXR"

, then we can get a new func via func sequence (or composing funcs):

	"acq_rel" : "acquire | release"

and mapping for xchg_relaxed():

	"%r = xchg_relaxed(%x, constvar:c);" -> "MOV %tmp,&c;
						 loop:load:LDXR %r,[%x];
						 store:STXR w%wmp, %tmp,[%x];
						 CBNZ w%wmp, loop"

, when we apply this func:

	"%r = xchg_acq_rel(%x, constvar:c);" -> "@acq_rel %r = xchg_relaxed(%x, constvar:c);"

, we expand "acq_rel" as func sequence:

	"%r = xchg_acq_rel(%x, constvar:c);" -> "@(acquire | release" %r = xchg_relaxed(%x, constvar:c);"

, apply "acquire" first:

	"%r = xchg_acq_rel(%x, constvar:c);" -> "@release %r = tmp_xchg_acquire(%x, constvar:c);"

	where we have a unobservable tmp rule:

	"%r = tmp_xchg_acquire(%x, constvar:c);" -> "MOV %tmp,&c;
						 loop:load:LDAXR %r,[%x];
						 store:STXR w%wmp, %tmp,[%x];
						 CBNZ w%wmp, loop"
, and apply "release":

	"%r = xchg_acq_rel(%x, constvar:c);" -> "MOV %tmp,&c;
						 loop:load:LDAXR %r,[%x];
						 store:STXLR w%wmp, %tmp,[%x];
						 CBNZ w%wmp, loop"

With this ability, we can save a few lines in the call file, because we
don't need to write intermediate callsites to generate intermediate
rules, e.g. temp_spin_lock() in Linux.call.

Also, while I'm at it, clean a bit of the code.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>