[herd]: Added LDR (literal), loads words from label
This commit adds support for AArch64 load literal from label.
This instruction is rather unusual in that it infers an address
from a label and loads a word from that address. In practice
we may see something like:

ldr X0, label_for_x

which is known by a runtime program, in practice such labels are
often emitted by the compiler with data aligned to section names
such as _DYNAMIC+0xf00 which corresponds to some other address. L230000
(for some hex 230000).

This convention is compiler specific, and up to the toolchain +
disassembler to correctly emit the correct construct. As such it is
safer to rely on the concrete address 230000 and parse it from a label:

ldr X0, L230000

parses 0x230000 from the label, and if we know &x=0x230000 then we can get
x. An example of this is in A17.litmus test