[all]: Handled Code Review Comments
This handles commits for the PR:
https://github.com/herd/herdtools7/pull/30

This commit makes the following changes:
- changed how parse_lbl parses to handle numeric addresses without
Str or the L prefix. For example:
LDR X0, 0x23000 // parses the decimal integer 0x230000
LDR X0, 230000  // parses the hexadecimal integer 23000
LDR X0, x       // parses the variable x

We should be careful with these kinds of literal instructions as they
are strictly not part of the Arm arm, but a convenience which operates
in an equivalent fashion.

- fixed ldr literal semantics to load based on size of register
- Removed dependency on Str entirely
- Changed RSpecial to SP in AArch64Base
- Changed barrel shifter type from 'k s option to 'k s
where s has a unit constructor NOEXT
- cleaned up rest of smaller issues in PR.