Telechat (#30)
* [all]: LDP/STP Support + concrete numeric addresses

This patch upstreams support for the AArch64 LDP/STP instructions.
This also upstreams support for providing numeric addresses in the
initial state. The patch contributions are outlined as follows
with motivations:

- Upstream LDP/STP support
- Added new address operator to the initial state
    you can write &x = 0xf00, then when a numeric address
    is passed to ldp/stp, it can dereference this address to
    get x. Later patches will upstream support for many more
    data access instructions like this. This kind of code is
    frequently emitted by compilers.
- Added address_of and deref unops to eventsMonad, this allows
  the solver to move between concrete '0xf00' and symbolic
  representations 'x' of addresses so that we may increment addresses,
  such as x+64 in ldp/stp, or acquire a symbolic address in the above
  example.
- This patch is supported by multiple unittests in herd/unittests to
  be used as regressions, see herd/unittests/README for instructions
  on what this means. As a policy any new instruction/variant should
  be supported with unittests and should be rejected without.

Some notes:
- Passing in symbolic variables still works as expected, this patch
  simply increases the kinds of input to the instructions we support
- Later patches will upstream support for many more lx**p, st**p, ld**r,
  st**r instructions, as well as pre-index, and barrel-shift
instructions.

* [lib]: HINTs as Nops in AArch64

This patch upstreams support for the HINT instruction. This instruction
has specific meaning allocated on a case by case basis, or otherwise

implemented as a NOP, as noted here:

https://developer.arm.com/docs/ddi0596/e/base-instructions-alphabetic-order/hint-hint-instruction

Change-Id: I15c3d7b408eca54be906b233fa8f689ef745354e

* [.gitignore]: added swp to gitignore

Change-Id: Ibe59822b920f8e5202e1153208c9a2add3ac6430

* [gen]: Added addret flag to prepend voidstar type

This allows diy to generate valid C++11 code.
The use of voidstar makes the thread functions valid
C++11 so that they can be compiled and used as a concurrent
program.

For instance, running

diy7 -arch C -addret

will generate C functions with 'void \*' preprended on front of each.

Change-Id: I3e2e101b93a8f494d786fec17e28bc4fef5e6161

* [lib]: Added SP as a special register to AArch64Base

Change-Id: I55bb2ae780af60559cca9ec6341279da842ef7a8

* [lib]: Made check_name public for use in other menhir Lexers

Makes use of the partial specification feature so we can reuse
the lexing machinery of AArch64 in other places

Change-Id: Id0738f4adc54624ca7b91ffb8e2687221f020f36

* [lib]: Added uniq_lists to MySet

handy function that will be handy later

* [tools]: Added '\_' to support global naming in mcompare

This allows us to parse more interesting names with underscore in

* [herd]: Added LDR (literal), loads words from label

This commit adds support for AArch64 load literal from label.
This instruction is rather unusual in that it infers an address
from a label and loads a word from that address. In practice
we may see something like:

ldr X0, label_for_x

which is known by a runtime program, in practice such labels are
often emitted by the compiler with data aligned to section names
such as _DYNAMIC+0xf00 which corresponds to some other address. L230000
(for some hex 230000).

This convention is compiler specific, and up to the toolchain +
disassembler to correctly emit the correct construct. As such it is
safer to rely on the concrete address 230000 and parse it from a label:

ldr X0, L230000

parses 0x230000 from the label, and if we know &x=0x230000 then we can get
x. An example of this is in A17.litmus test

* [herd]: Added MOVZ operation with barrel shift operands

This patch upstreams support for the MOVZ operation in AArch64
with logical shift left (LSL)  operand:

movz Wn, #imm{, LSL #shift1}
movz Xn, #imm{, LSL #shift2}

where shift1 is 0 or 16 for 32-bit register
where shift2 is 0, 16, 32, 48 for 64-bit register

* [herd]: Added ADR,ADRP numeric and symbolic

     This commit adds support for AArch64 ADR and ADRP load address.
     This instruction is rather unusual in that it infers an address
     from a label. In practice we may see something like:

     adr X0, x
     adrp X0, x

     in practice such labels are often emitted by the compiler with
     data aligned to section names such as _DYNAMIC+0xf00 which
     corresponds to some numeric address 0x230000.
     We also support numeric addresses:

     adr X0, L230000
     adrp X0, L230000

     This convention is compiler specific, and up to the toolchain +
     disassembler to correctly emit the correct construct. As such it is
     safer to rely on the concrete address 230000 and parse it from a
label:

     adr X0, L230000

     parses 0x230000 from the label. Unittests included for both cases

* [all]: Handled Code Review Comments

This handles commits for the PR:
https://github.com/herd/herdtools7/pull/30

This commit makes the following changes:
- changed how parse_lbl parses to handle numeric addresses without
Str or the L prefix. For example:
LDR X0, 0x23000 // parses the decimal integer 0x230000
LDR X0, 230000  // parses the hexadecimal integer 23000
LDR X0, x       // parses the variable x

We should be careful with these kinds of literal instructions as they
are strictly not part of the Arm arm, but a convenience which operates
in an equivalent fashion.

- fixed ldr literal semantics to load based on size of register
- Removed dependency on Str entirely
- Changed RSpecial to SP in AArch64Base
- Changed barrel shifter type from 'k s option to 'k s
where s has a unit constructor NOEXT
- cleaned up rest of smaller issues in PR.

* [herd]: removed no-op in movz LSL, Added Failure catch