[herd]: -allow_numeric_literals flag for literal instrs
This patch adds a flag to allow literal instructions to accept numeric
values as operands in place of symbolic labels. For example we allow:

ADR Xn, x

to load the location x in to Xn, however it may be convenient to load
the numeric address of x into Xn (in the case we are checking
disassembly, for instance), this flag allows us to parse such as case:

ADR Xn, 230000 // for decimal value 230000
ADR Xn, 0x230000 for Hexadecimal value 0x23000

The same thus applies for load literal instructions. We should be
careful with this instruction, it is strictly not part of the arm arm.