[all]: Addressed Code review comments.
This is Patch 1/2 to address the first round of comments for
pull request https://github.com/herd/herdtools7/pull/31.

In detail this patch changes the following:
- Adds Test A95.litmus to handle add register with left shift immediate
- removes SignedExtendWord from lib/op, implements in in AArch64Sem
- removes arguments from S_SXTW/UXTW
- combines movk/movz where immediately possible
- Addressed comments about ROR
- fixed bugs/code style comments/errrors
- addressed comment about ASR shifter
- added match_shift function to jingle to handle shift comparisons

Another patch will address the remainder of the comments later, there
is enough here that might require further scrutiny before I cover the
rest.