[diy]: Added NoVolatile variant to diy
Added the option to generate C tests without the Volatile option.
Volatile is not currently used for any other reason than to prevent
compiler optimisations to be applied in litmus. If we are to use such
optimisations we must trust they are correct.

This patch removes the qualifier on types as it is not needed in this
case.

No unittests are provided but you can run the following:

diy7 -arch C -variant NoVolatile

and test with

herd7 -model rc11.cat *.litmus