[lib]: Added void ctype
Consider the C code fragment:
```
void *P0 (...) {
    ....
}
```
which is the (pthread friendly) format of concurrent C functions. We may
wish to generate this using diy, and so we would need to know to print a
void pointer type. This patch adds the `void` type in lib, which we can
later combine to generate such a code fragment.