[herd]: Added partial support for AArch64 LDP/STP
Following the review of https://github.com/herd/herdtools7/pull/29

This patch upstreams support for Load/Store pair instructions for
AArch64 in Herd. The patch linked above contains more code for numeric
addressing which has been removed here. In detail this patch does:

- partial support for Wreg load/store pairs, we cannot handle the xreg
  case just yet as it requires more work.

This patch does not:

- Add any kind of numeric addressing capabilities as alluded to in #29.
- Add any kind of support for vectors - or indeed a large 128-bit type.

This patch raises questions about what we do in the fully symbolic case
for:

STP Xn, Xm, [Xn] where Xn and Xm contain 64-bit values and Xn points to
a value

In herd the largest possible sized type is int64_t and so storing Xn
causes out-of-bounds errors as we are writing 128-bits worth of data. We
have a solution to this in the works, but for now we can safely upstream
this patch so that it works for wregs. It may be preferable to add a
pattern for 'unsupported xregs' until we upstream the solution for xregs.

The same reasoning applies to LDP instructions