[lib] Add morello
Add a number of morello instructions. Some of which are not particularly
useful at the moment given that addresses are represented as strings and
not numerical values.

This patch includes the implementation of the tag, permissions and
object type of capabilities. It doesn't include their bounds and flags.
When using morello, the only implemented instruction set is C64.

Add the necessary operations on Value to implement those instructions.

Add the possibility to use unary and binary operations on Symbolic.
Those operation will only make use of the new "capability" field of
Symbolic values, and won't have an effect on non-morello variants.