[gen] Add morello
Add capability annotations for tag "Ct" and seal "Cs". Those annotations
correspond to their own Banks, behaving similarly to Ord.

A situation of potential Fault is created when using those annotations
together with DpAddr. The result of the preceding read is used to modify
the address so that it will fault if different from the expected value.

Implement the addition of the optional annotation modifier for morello
capabilities.