[litmus] Set memory attributes in vars_init(...)
In AArch64 accessing a location with different cacheability attributes
can lead to loss of coherence. This patch ensures that initial writes
and cache maintenance prior to the code of the test are performed
after setting the relevant PTE. In addition, and to make sure that
there are no stale copies in the cache, we clean and invalidate cache
copies.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>