[herd] KVM,AArch64 Use test-and-set for updating the AF and DB bits, (#85)
This PR features a significant change in HW management of the access and dirty bits. Those are now treated independently by test-and-set operation on pte's. This provides more behaviours than the previous monolithic approach, at a certain price in running time. The reason of this lower performance is the numerous choices introduced by test-and-set operations.
+ AF update is attempted unconditionally and performed when
    valid bit is set and AF unset
+ DB update is attempted when appropriate (DB unset and write
    decided to be performed) and succeeds when DB unset.