[diy]: Add NoVolatile value to Variant.t in diy
This patch adds the NoVolatile enum to Variant.t which will configure
whether to remove the volatile qualifier on types in C generated tests.

Volatile is not currently used for any other reason than to prevent
compiler optimisations being applied in litmus. If we are to use such
optimisations we must trust and test that they are correct.

No unittests are provided yet - infrastructure to test this is in the
works. For now you can run test the implementation by running
```
diy7 -arch C -variant NoVolatile
```
and check the output C tests do not have volatile qualified types,
you can check these are then accepted by Herd using
```
herd7 -model rc11.cat *.litmus
```
We can follow this up with tests once the infrastructure to do this exists.