[all]: Added vector ADT and parsing in initial state.
This patch upstreams support for the sized-bounded vector type
in Herd, jingle, etc... unittests are included for completeness.

Specifically this patch adds parser support for vectors in the initial
state of litmus tests, and a corresponding type underneath so we can
write e.g:

```
int64_t v[2] = {1, 2}
```
which is expanded to v[0] = 1; v[1] = 2; etc...

Checks are added in the parser and elaboration stage in order
to ensure the size matches the supplied arguments.

Open questions (to fix before we merge):
- There is a conflict involving the syntax and semantics of
  the vector type. See test A130.litmus for a full explanation

in short: there is a type in herd that uses C list syntax we want:
```
v[8] = 3;    (* this is v+8 bytes*)
```
however the semantics we need is used by another type:
```
v+1 = 3; (* offset by the size of the type, so when uint64_t v, v+1 is
the same as v+8 bytes*)
```
ie we want the syntax of the first one but the semantics of the second.
These are 2 different types in Herd.

One Solution: We accept the first form so we have the C-list syntax, but
all offsets are byte offsets, and not the size of the type. E.g.
```
uint64 v[4] = { 1,2,3,4 }
P0;
(* program ...*)
forall (v[8] = 2) (*rather than v[1] = 2 *)
```
- There is an issue with `is_non_mixed_symbol` which checks the
  alignment of symbolic values, this fails in the presence of vectors
which are elaborated to v+0, v+8 etc... - it checks only that the
alignment is with respect to one element. e.g v+8 fails to align to the
second element of uint64_t v[2] = { 1,2 } since it checks the size of
uint64_t v.

I have made this obvious during the build process with a Warning, but it
would be worth getting your thoughts as this affects -variant unaligned.
etc...

- There is a shift reduce conflict in the parser owing to the
  introduction of {} into the initial state. I had a look at
  this but need to fix this - any tips?

Co-Authored-By: Luc Maranget