[internal] Add function Test.fail for use in unit tests
Because Failure & failwith are built-ins, I want to avoid using them for
test assertion failures, which might catch / mask / mis-represent
Failure exceptions elsewhere in the callstack.

A contrived example:

  let f xs =
    match xs with
    | [] -> failwith "this should never happen"
    | xs -> ...

  let g xs =
    (* uses f in some way, possibly deep in its callstack. *)

  . . .

  (* xs has become []. *)
  if g xs <> 0 then
    failwith "Did not return 0"

This will print as:

  Failed Foo.g: this should never happen

It should now print as:

  Failed Foo.g: raised exception
  Fatal error: exception Failure("this should never happen")

I have called the exception AssertionFailure and the function Test.fail
to ensure that they won't pun the built-in Failure & failwith.