[herd]: Fix same-register I_OP3 barrel shift bug
Consider the following:
```
ADD W1, W0, W0, lsl #1;
```
when `0:X0=1`

the final result `0:X1=3` should always occur,
however `0:X1=4` occurs since a single variable `v` is used
when the registers are the same, when two variables are required.

This patch fixes the above for `I_OP3` and adds a `MOVZ` test
missed before.