Variant in test metadata (#109)
This PR introduces "variants" in the metadata of tests. This is done by using a "Variant" field.

Writing
```
Variant=<tags>
```
Has the same effect as adding option `-variant <tags>` at the end the the command line. As a consequence, variants from the test take precedence over the variants from the command line. 

* [lib] Share code for parsing the -variant option.
  Code is made slightly more complex because of
  variants that are to take effect at parsing time.
  The only such variant at the moment is precise/imprecise that
  sets/clears the `precision` boolean configuration field in herd.

* [herd] Handle the variant field of test meta-data.

* [litmus] Handle variant in test meta-data.