[tools] Handle old log files (_i.e._ without explicit absent faults) in sums.
The idea is considering that outcomes that differ only in fault absence are
equal. Then, the outputed oucome is the one with explicit fault absences.

Minimal safety check is added to the `compare_state`, by checking that
outcomes that differ only by the presence of fault absence do not occur
in the outcomes of a given test in a given log file.