[lib] Add support for memory attributes in PTEVal
Add a new field attrs in the PTEVal struct. This field a free-form
list of strings and the specific values will depend on the
architecture. In addition, this change adds support for attributes in
the litmus tests parser. The PTE attributes are parsed as coma
separated strings in the pte.

For example:

pte_x=(oa:phy_y,attrs:(Normal,Inner-write-back,Outer-write-back,Inner-shareable));

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>