[all]: Add vector metadata to Symbolic type
This patch addresses feedback from #94

This patch is not quite ready yet, as mixed-mode for vectors does
not work, and we have a new form of test for 'deliberate errors'
(see A134 test)

As this commit is quite large so I have split it into first patch and
second.

In detail this patch:
- Rebase and rename tests in testsuite - all pass
- run `make versions` to check backwards compatibility - passes with no
  warnings
- applied Luc's obsolete syntax patch to fix the shift-reduce error in
  the parser
- modifies tests so they  use `v[n] = {1,2,3}` `v[0] = 1; v[1] = 2 ...` syntax
- remove unallocated arrays in tests, make test A134 deliberately fails with
unallocated array error. The test infrastructure can not handle expected failures yet.
I have spoken to eth-arm about this and will follow up this patch with a fix.
- Updated A08 test with new `v[n]` syntax.
- Add optional vector metadata to Symbolic type including size of primitive type,
and total-size of vector. The initial state constructs the metadata, and we use
it in the final state (see lib/constant). Alignment checks are added for
vector elements.

Co-Authored-By: Luc Maranget