[internal_lib] Unify Command.run_with_* into Command.run
The number of functions to run commands has grown to 3:
- Command.run,
- Command.run_with_stdout,
- Command.run_with_stdin_and_stdout.

As upcoming PR from someone else is going to need to use the stderr of a
command, this is going to grow to 4, and may eventually have one for
each of the full powerset of {stdin,stdout,stderr}.

Instead, this commit implements One True Function, Command.run, which
takes 3 optional function parameters, one for each of ~stdin, ~stdout,
and ~stderr.

The existing Command.run_with_* functions are still in place,
implemented in terms of the unified Command.run, and will be cleaned up
in a later commit.