Support for expected-failure tests
This patch unblocks the expected failures in #94
This patch depends on changes in #132 and should not be merged before

Consider the test `LB005.litmus`:
```
fictional_instruction X0, X1;
```
which should fail in Herd with a parser error, written in
`LB005.litmus.expected-failure`
```
Warning: File "./herd/tests/instructions/AArch64/L005.litmus", line 7, characters 22-24: unexpected 'X0' (in prog) (User error)
```
This patch extends the testing infrastructure to test for expected failures.
This unblocks #94, in particular allows us to test an expected failure of Herd.