Added a 'loc rloc type.
This type represents the content of locations (rvalue) in two situations.
  + 'locations' in condition atoms loc=v
  + locations in the 'locations' field.

This `'loc rval` is defined as either a location (`Loc of 'loc) or an
array access (`Deref of 'loc * int`).

This is still perhaps too general
since we are only interested in
  + register locations
  + global locations, without offset, nor other decorations such as tags.
  + array accesses, where the array is a global location as above.

For instance, we do not need array accesses to location registers. However
some additional locations maybe useful in the future (such as page table
entry locations). So, let us try it that way.