Add el0 field to page table entry value (pteval). (#136)
This field ---AP[1], bit 6 of pteval's--- controls access to virtual memory from exception level EL0. A value of 0 forbids access, a value of 1 allows access.

It is to be noticed that adding a field to pteval's would result in untimely consequences if one does not change the printing of pteval's (print all fields). To avoid those, this PR also includes the following changes.

Output by all tools is now changed, fields are not printed when their value is default.
For handling old logs, pteval's are normalised to new format internally by tools that consume logs (mcompare, msum..).
While computing hashes, a special printing mode is used: old mode for old fields (always print) and new mode for new fields (print when non-default).

* [herd] Terser printing for pteval's.
In practice, fields whose value is default are elided.
The old behaviour (print all fields) is kept for computing
test hashes.

* [litmus] Terser litmus printing of pteval's in outcomes.

* [tools] Normalise pteval's from logs.
  Normalisation is necessary to combine and compare old and new logs.
  Namely, pteval's in old logs have all fields explicit, while fields
  with default values are omitted in new logs.

* [lib] Add el0 field in definition of pteval's.
  The el0 field controls access to memory from level EL0.
  Values: 1 (default) -> allowed, 0 -> forbidden.

* [litmus] Implement el0 field.

* [herd] Implement el0 field.