Vector reloaded (#140)
* [all]: Added vector ADT and parsing in initial state.

This patch upstreams support for the sized-bounded vector type
in Herd, jingle, etc... unittests are included for completeness.

Specifically this patch adds parser support for vectors in the initial
state of litmus tests, and a corresponding type underneath so we can
write e.g:

```
int64_t v[2] = {1, 2}
```
which is expanded to v[0] = 1; v[1] = 2; etc...

Checks are added in the parser and elaboration stage in order
to ensure the size matches the supplied arguments.

Open questions (to fix before we merge):
- There is a conflict involving the syntax and semantics of
  the vector type. See test A130.litmus for a full explanation

in short: there is a type in herd that uses C list syntax we want:
```
v[8] = 3;    (* this is v+8 bytes*)
```
however the semantics we need is used by another type:
```
v+1 = 3; (* offset by the size of the type, so when uint64_t v, v+1 is
the same as v+8 bytes*)
```
ie we want the syntax of the first one but the semantics of the second.
These are 2 different types in Herd.

One Solution: We accept the first form so we have the C-list syntax, but
all offsets are byte offsets, and not the size of the type. E.g.
```
uint64 v[4] = { 1,2,3,4 }
P0;
(* program ...*)
forall (v[8] = 2) (*rather than v[1] = 2 *)
```
- There is an issue with `is_non_mixed_symbol` which checks the
  alignment of symbolic values, this fails in the presence of vectors
which are elaborated to v+0, v+8 etc... - it checks only that the
alignment is with respect to one element. e.g v+8 fails to align to the
second element of uint64_t v[2] = { 1,2 } since it checks the size of
uint64_t v.

I have made this obvious during the build process with a Warning, but it
would be worth getting your thoughts as this affects -variant unaligned.
etc...

- There is a shift reduce conflict in the parser owing to the
  introduction of {} into the initial state. I had a look at
  this but need to fix this - any tips?

Co-Authored-By: Luc Maranget

* [all]: Add vector metadata to Symbolic type

This patch addresses feedback from #94

This patch is not quite ready yet, as mixed-mode for vectors does
not work, and we have a new form of test for 'deliberate errors'
(see A134 test)

As this commit is quite large so I have split it into first patch and
second.

In detail this patch:
- Rebase and rename tests in testsuite - all pass
- run `make versions` to check backwards compatibility - passes with no
  warnings
- applied Luc's obsolete syntax patch to fix the shift-reduce error in
  the parser
- modifies tests so they  use `v[n] = {1,2,3}` `v[0] = 1; v[1] = 2 ...` syntax
- remove unallocated arrays in tests, make test A134 deliberately fails with
unallocated array error. The test infrastructure can not handle expected failures yet.
I have spoken to eth-arm about this and will follow up this patch with a fix.
- Updated A08 test with new `v[n]` syntax.
- Add optional vector metadata to Symbolic type including size of primitive type,
and total-size of vector. The initial state constructs the metadata, and we use
it in the final state (see lib/constant). Alignment checks are added for
vector elements.

Co-Authored-By: Luc Maranget

* Direct print of symbols.

* Supress constructor Location_deref. + Too invasive + Superseeded by offsets in symbols.

* Added a 'loc rloc type.

This type represents the content of locations (rvalue) in two situations.
  + 'locations' in condition atoms loc=v
  + locations in the 'locations' field.

This `'loc rval` is defined as either a location (`Loc of 'loc) or an
array access (`Deref of 'loc * int`).

This is still perhaps too general
since we are only interested in
  + register locations
  + global locations, without offset, nor other decorations such as tags.
  + array accesses, where the array is a global location as above.

For instance, we do not need array accesses to location registers. However
some additional locations maybe useful in the future (such as page table
entry locations). So, let us try it that way.

* [herd/lib]: Use Deref to implement scaled accesses

This patch does the following:

- Prohibit `v[i] = scalar` syntax in init, allow only
  `v[i]={scalar_list}` syntax. Test A175 added
- Change printing of Symbolic type so it is v+i for internal accesses,
  and v[i] (scaled) for external/user-facing view
- rename A134.litmus.expected to A134.litmus.expected-failure in line
  with #133

* [herd/lib]: Address Code review comments

This patch addresses the code review comments for the pull request
but does not handle 2 comments which will be handled in the following
commit

* [all]: Change Symbolic type to use record

Before this patch, the `Symbolic` type in `lib/constant`
had type:
```
Symbolic of (string * tag * cap * vdata) * offset
```
which is quite complex, this patch changes the type:
```
Symbolic of symbolic_data
```
where `symbolic_data` contains the data in a record.

:x

* [herd/tests]: Add relative path of tests to expected failure msg

* [herd]: Added litmus tests for vectors

Luc Maranget kindly provided additional tests `L006`, `L007`, and `L008` which I have
added to the test suite, and fixed `lib/constant` so that all tests pass

* [herd] Add (correct) size component to vector element initial writes

Notice that the size of one element is associated
to the vector location in size environment.

* [herd] Add type environment to test structure.

This type environment is first used to scale array references in
final states. It can also be used to correct wrong printing of
small size negative integer represented as Int64 integers.

* [herd] Use typing to scale offsets in final states.

This paves the way for a type directing output of final states.

Co-authored-by: lukeg101 <6547672+lukeg101@users.noreply.github.com>